renamed the directories
